fix(server): disable artist name search
