+++ src/blueprints/api/v1/blueprint.py
@@ -68,10 +68,13 @@ def list_artists(page: str):
 
     search_params: TDArtistsParams = result['data']
     service = search_params["service"]
-    artist_name = search_params["name"]
+    # artist_name = search_params["name"]
     current_page = parse_int(page)
     limit = DEFAULT_PAGE_LIMIT
-    artist_count = count_artists(service, artist_name)
+    artist_count = count_artists(
+        service,
+        # artist_name
+    )
     total_pages = math.floor(artist_count / limit) + 1
     is_valid_page = current_page and current_page <= total_pages
 
@@ -96,7 +99,11 @@ def list_artists(page: str):
         offset=offset,
         sql_limit=sql_limit
     )
-    artists = get_artists(pagination_db, service, artist_name)
+    artists = get_artists(
+        pagination_db,
+        service,
+        # artist_name
+    )
     pagination = TDPagination(
         total_count=artist_count,
         total_pages=total_pages,
